+++ Libraries/Fields/Image.php
@@ -64,13 +64,22 @@ class Image extends Field {
 	public function __construct($field, $info, $config)
 	{
 		parent::__construct($field, $info, $config);
+		$isSettings = is_a($config, 'Admin\\Libraries\\SettingsConfig');
 
 		$this->sizes = array_get($info, 'sizes', $this->sizes);
 		$this->naming = array_get($info, 'naming', $this->naming);
 		$this->length = array_get($info, 'length', $this->length);
 		$this->location = array_get($info, 'location');
 		$this->sizeLimit = (int) array_get($info, 'size_limit', $this->sizeLimit);
-		$this->uploadUrl = URL::to_route('admin_image_upload', array($config->name, $this->field));
+
+		if ($isSettings)
+		{
+			$this->uploadUrl = URL::to_route('admin_settings_image_upload', array($config->name, $this->field));
+		}
+		else
+		{
+			$this->uploadUrl = URL::to_route('admin_image_upload', array($config->name, $this->field));
+		}
 
 		//make sure the naming is one of the two accepted values
 		$this->naming = in_array($this->naming, array('keep', 'random')) ? $this->naming : 'random';